Reworked testing more
useful comment
