+++ tests/test_friend_search.py
